fix: moved lighting action method to lightning store
